refactor: split speculative philosophy extension
