+++ src/extensions/speculative-philosophy/citation.ts
@@ -1,0 +1,249 @@
+import { ToolDefinition, ExecuteToolResponseSchema } from "@anthropic-ai/mcp-typescript-sdk";
+export const PHILOSOPHICAL_CITATION_TOOL: ToolDefinition = {
+  name: 'format-philosophical-citation',
+  description: 'Formats citations according to philosophical tradition standards',
+  parameters: {
+    type: 'object',
+    properties: {
+      sourceType: {
+        type: 'string',
+        enum: ['book', 'article', 'anthology', 'primary-text', 'translation', 'lecture'],
+        description: 'Type of source being cited',
+      },
+      citationDetails: {
+        type: 'object',
+        properties: {
+          author: { type: 'string' },
+          title: { type: 'string' },
+          year: { type: 'string' },
+          publisher: { type: 'string' },
+          translator: { type: 'string' },
+          editor: { type: 'string' },
+          journal: { type: 'string' },
+          volume: { type: 'string' },
+          pages: { type: 'string' },
+          edition: { type: 'string' },
+          url: { type: 'string' },
+        },
+        description: 'Details of the citation',
+      },
+      citationStyle: {
+        type: 'string',
+        enum: ['chicago', 'mla', 'apa', 'oxford', 'harvard', 'custom'],
+        description: 'Style guide to follow',
+        default: 'chicago',
+      },
+      inlineCitation: {
+        type: 'boolean',
+        description: 'Whether to generate both inline citation and bibliography entry',
+        default: true,
+      },
+    },
+    required: ['sourceType', 'citationDetails'],
+  },
+  returns: {
+    type: 'object',
+    properties: {
+      latex: {
+        type: 'string',
+        description: 'Generated LaTeX code for the citation',
+      },
+      inlineVersion: {
+        type: 'string',
+        description: 'Inline citation format',
+      },
+      bibliographyVersion: {
+        type: 'string',
+        description: 'Bibliography entry format',
+      },
+    },
+  },
+};
+
+/**
+ * Handles worldbuilding document creation
+ */
+async function handleWorldbuildingDocument(parameters: any): Promise<ExecuteToolResponseSchema> {
+  const { worldName, worldType = 'other', keyElements = [], description } = parameters;
+  
+  // This is where we'd integrate with Claude for advanced generation
+  // For now, we'll implement a template-based approach
+  
+  // Default sections for worldbuilding
+  const defaultSections = [
+    'Overview',
+    'History',
+    'Geography',
+    'Culture',
+    'Magic/Technology',
+    'Political Structure',
+export async function handlePhilosophicalCitation(parameters: any): Promise<ExecuteToolResponseSchema> {
+  const { 
+    sourceType, 
+    citationDetails,
+    citationStyle = 'chicago',
+    inlineCitation = true
+  } = parameters;
+  
+  // Extract citation details with defaults
+  const {
+    author = 'Author Name',
+    title = 'Title of Work',
+    year = '2023',
+    publisher = 'Publisher Name',
+    translator = '',
+    editor = '',
+    journal = '',
+    volume = '',
+    pages = '',
+    edition = '',
+    url = ''
+  } = citationDetails;
+  
+  // Generate citation based on style and source type
+  let inlineVersion = '';
+  let bibliographyVersion = '';
+  
+  if (citationStyle === 'chicago') {
+    // Chicago style citations
+    if (sourceType === 'book') {
+      inlineVersion = `(${author} ${year})`;
+      bibliographyVersion = `${author}. \\textit{${title}}${translator ? `, translated by ${translator}` : ''}${editor ? `, edited by ${editor}` : ''}${edition ? `, ${edition} ed.` : ''}. ${publisher}, ${year}.`;
+    } else if (sourceType === 'article') {
+      inlineVersion = `(${author} ${year})`;
+      bibliographyVersion = `${author}. ``${title}.'' \\textit{${journal}}${volume ? ` ${volume}` : ''}${pages ? `: ${pages}` : ''}, ${year}.`;
+    } else if (sourceType === 'primary-text') {
+      // For philosophical primary texts, often cited by section rather than page
+      inlineVersion = `(${author}, ${title.split(' ')[0]}, ${pages || 'section'})`;
+      bibliographyVersion = `${author}. \\textit{${title}}${translator ? `, translated by ${translator}` : ''}${editor ? `, edited by ${editor}` : ''}. ${publisher}, ${year}.`;
+    } else {
+      // Default format
+      inlineVersion = `(${author} ${year})`;
+      bibliographyVersion = `${author}. \\textit{${title}}. ${publisher}, ${year}.`;
+    }
+  } else if (citationStyle === 'apa') {
+    // APA style citations
+    if (sourceType === 'book') {
+      inlineVersion = `(${author}, ${year})`;
+      bibliographyVersion = `${author}. (${year}). \\textit{${title}}${edition ? ` (${edition} ed.)` : ''}${translator ? ` (${translator}, Trans.)` : ''}. ${publisher}.`;
+    } else if (sourceType === 'article') {
+      inlineVersion = `(${author}, ${year})`;
+      bibliographyVersion = `${author}. (${year}). ${title}. \\textit{${journal}}${volume ? `, ${volume}` : ''}${pages ? `, ${pages}` : ''}.`;
+    } else {
+      // Default format
+      inlineVersion = `(${author}, ${year})`;
+      bibliographyVersion = `${author}. (${year}). \\textit{${title}}. ${publisher}.`;
+    }
+  } else {
+    // Default to MLA style
+    if (sourceType === 'book') {
+      inlineVersion = `(${author} ${pages})`;
+      bibliographyVersion = `${author}. \\textit{${title}}${translator ? `, translated by ${translator}` : ''}${editor ? `, edited by ${editor}` : ''}. ${publisher}, ${year}.`;
+    } else if (sourceType === 'article') {
+      inlineVersion = `(${author} ${pages})`;
+      bibliographyVersion = `${author}. ``${title}.'' \\textit{${journal}}${volume ? `, vol. ${volume}` : ''}${pages ? `, ${pages}` : ''}, ${year}.`;
+    } else {
+      // Default format
+      inlineVersion = `(${author} ${pages || year})`;
+      bibliographyVersion = `${author}. \\textit{${title}}. ${publisher}, ${year}.`;
+    }
+  }
+  
+  // Generate BibTeX entry
+  let bibtexEntry = '';
+  const authorLastName = author.split(' ').pop() || 'Author';
+  const citeKey = `${authorLastName.toLowerCase()}${year}`;
+  
+  if (sourceType === 'book') {
+    bibtexEntry = `@book{${citeKey},
+  author = {${author}},
+  title = {${title}},
+  year = {${year}},
+  publisher = {${publisher}}${edition ? `,\n  edition = {${edition}}` : ''}${translator ? `,\n  translator = {${translator}}` : ''}${editor ? `,\n  editor = {${editor}}` : ''}
+}`;
+  } else if (sourceType === 'article') {
+    bibtexEntry = `@article{${citeKey},
+  author = {${author}},
+  title = {${title}},
+  journal = {${journal}}${volume ? `,\n  volume = {${volume}}` : ''}${pages ? `,\n  pages = {${pages}}` : ''},
+  year = {${year}}
+}`;
+  } else if (sourceType === 'primary-text') {
+    bibtexEntry = `@book{${citeKey},
+  author = {${author}},
+  title = {${title}},
+  year = {${year}},
+  publisher = {${publisher}}${translator ? `,\n  translator = {${translator}}` : ''}${editor ? `,\n  editor = {${editor}}` : ''}
+}`;
+  } else {
+    bibtexEntry = `@misc{${citeKey},
+  author = {${author}},
+  title = {${title}},
+  year = {${year}}${publisher ? `,\n  publisher = {${publisher}}` : ''}${url ? `,\n  url = {${url}}` : ''}
+}`;
+  }
+  
+  // Generate LaTeX document with citation examples
+  const latex = `\\documentclass{article}
+\\usepackage{geometry}
+\\usepackage{xcolor}
+\\usepackage{hyperref}
+\\usepackage{natbib}  % For citation management
+
+\\title{Philosophical Citation Example}
+\\author{}
+\\date{}
+
+\\begin{document}
+
+\\maketitle
+
+\\section{Citation Example}
+
+\\subsection{Source Information}
+\\begin{itemize}
+  \\item Source Type: ${sourceType}
+  \\item Author: ${author}
+  \\item Title: ${title}
+  \\item Year: ${year}
+  ${publisher ? `\\item Publisher: ${publisher}` : ''}
+  ${journal ? `\\item Journal: ${journal}` : ''}
+  ${volume ? `\\item Volume: ${volume}` : ''}
+  ${pages ? `\\item Pages: ${pages}` : ''}
+  ${translator ? `\\item Translator: ${translator}` : ''}
+  ${editor ? `\\item Editor: ${editor}` : ''}
+  ${edition ? `\\item Edition: ${edition}` : ''}
+  ${url ? `\\item URL: ${url}` : ''}
+\\end{itemize}
+
+\\subsection{${citationStyle.toUpperCase()} Style Citations}
+
+\\subsubsection{In-text Citation}
+${inlineCitation ? inlineVersion : 'No in-text citation requested.'}
+
+\\subsubsection{Bibliography Entry}
+${bibliographyVersion}
+
+\\subsection{BibTeX Entry}
+\\begin{verbatim}
+${bibtexEntry}
+\\end{verbatim}
+
+\\subsection{Usage Example}
+
+According to philosophical tradition ${inlineVersion}, the concept of being is fundamental to metaphysical inquiry. This perspective has been influential in continental philosophy.
+
+\\begin{thebibliography}{9}
+\\bibitem{${citeKey}} ${bibliographyVersion}
+\\end{thebibliography}
+
+\\end{document}`;
+
+  return {
+    result: {
+      latex,
+      inlineVersion,
+      bibliographyVersion
+    }
+  };
+}
