+++ src/extensions/speculative-philosophy/index.ts
@@ -9,61 +9,14 @@ import {
 } from '@anthropic-ai/mcp-typescript-sdk';
 import { Claude } from '@anthropic/sdk';
 
+import { WORLDBUILDING_TOOL, handleWorldbuildingDocument } from "./worldbuilding";
+import { CHARACTER_DEVELOPMENT_TOOL, handleCharacterProfile } from "./character";
+import { PHILOSOPHICAL_CITATION_TOOL, handlePhilosophicalCitation } from "./citation";
 // Initialize Claude client for advanced generation capabilities
 const claude = new Claude({
   apiKey: process.env.ANTHROPIC_API_KEY
 });
 
-/**
- * Speculative Fiction Worldbuilding Tool
- * Helps create structured LaTeX documentation for fictional worlds
- */
-const WORLDBUILDING_TOOL: ToolDefinition = {
-  name: 'create-worldbuilding-document',
-  description: 'Creates structured LaTeX documentation for speculative fiction worldbuilding',
-  parameters: {
-    type: 'object',
-    properties: {
-      worldName: {
-        type: 'string',
-        description: 'Name of the fictional world',
-      },
-      worldType: {
-        type: 'string',
-        enum: ['sci-fi', 'fantasy', 'alternate-history', 'post-apocalyptic', 'magical-realism', 'cyberpunk', 'solarpunk', 'other'],
-        description: 'Type of speculative fiction world',
-      },
-      keyElements: {
-        type: 'array',
-        items: {
-          type: 'string',
-        },
-        description: 'Key world elements to document (e.g., magic system, technology, history)',
-      },
-      description: {
-        type: 'string',
-        description: 'Brief description of the world and its core concepts',
-      },
-    },
-    required: ['worldName', 'description'],
-  },
-  returns: {
-    type: 'object',
-    properties: {
-      latex: {
-        type: 'string',
-        description: 'Generated LaTeX code for the worldbuilding document',
-      },
-      sections: {
-        type: 'array',
-        items: {
-          type: 'string',
-        },
-        description: 'List of sections created in the document',
-      },
-    },
-  },
-};
 
 /**
  * Philosophical Argument Structure Tool
@@ -118,57 +71,6 @@ const PHILOSOPHICAL_ARGUMENT_TOOL: ToolDefinition = {
 };
 
 /**
- * Character Development Tool
- * Creates LaTeX templates for character development in speculative fiction
- */
-const CHARACTER_DEVELOPMENT_TOOL: ToolDefinition = {
-  name: 'create-character-profile',
-  description: 'Creates LaTeX templates for detailed character profiles in speculative fiction',
-  parameters: {
-    type: 'object',
-    properties: {
-      characterName: {
-        type: 'string',
-        description: 'Name of the character',
-      },
-      characterType: {
-        type: 'string',
-        enum: ['protagonist', 'antagonist', 'supporting', 'mentor', 'foil', 'other'],
-        description: 'Role of the character in the narrative',
-      },
-      basicDescription: {
-        type: 'string',
-        description: 'Basic description of the character',
-      },
-      worldContext: {
-        type: 'string',
-        description: 'The speculative fiction context the character exists in',
-      },
-      detailLevel: {
-        type: 'string',
-        enum: ['basic', 'detailed', 'comprehensive'],
-        description: 'Level of detail for the character profile',
-        default: 'detailed',
-      },
-    },
-    required: ['characterName', 'basicDescription'],
-  },
-  returns: {
-    type: 'object',
-    properties: {
-      latex: {
-        type: 'string',
-        description: 'Generated LaTeX code for the character profile',
-      },
-      sections: {
-        type: 'array',
-        items: {
-          type: 'string',
-        },
-        description: 'List of sections in the character profile',
-      },
-    },
-  },
 };
 
 /**
@@ -349,213 +251,8 @@ const TIMELINE_TOOL: ToolDefinition = {
 };
 
 /**
- * Philosophical Citation Style Tool
- * Formats citations according to philosophical tradition standards
- */
-const PHILOSOPHICAL_CITATION_TOOL: ToolDefinition = {
-  name: 'format-philosophical-citation',
-  description: 'Formats citations according to philosophical tradition standards',
-  parameters: {
-    type: 'object',
-    properties: {
-      sourceType: {
-        type: 'string',
-        enum: ['book', 'article', 'anthology', 'primary-text', 'translation', 'lecture'],
-        description: 'Type of source being cited',
-      },
-      citationDetails: {
-        type: 'object',
-        properties: {
-          author: { type: 'string' },
-          title: { type: 'string' },
-          year: { type: 'string' },
-          publisher: { type: 'string' },
-          translator: { type: 'string' },
-          editor: { type: 'string' },
-          journal: { type: 'string' },
-          volume: { type: 'string' },
-          pages: { type: 'string' },
-          edition: { type: 'string' },
-          url: { type: 'string' },
-        },
-        description: 'Details of the citation',
-      },
-      citationStyle: {
-        type: 'string',
-        enum: ['chicago', 'mla', 'apa', 'oxford', 'harvard', 'custom'],
-        description: 'Style guide to follow',
-        default: 'chicago',
-      },
-      inlineCitation: {
-        type: 'boolean',
-        description: 'Whether to generate both inline citation and bibliography entry',
-        default: true,
-      },
-    },
-    required: ['sourceType', 'citationDetails'],
-  },
-  returns: {
-    type: 'object',
-    properties: {
-      latex: {
-        type: 'string',
-        description: 'Generated LaTeX code for the citation',
-      },
-      inlineVersion: {
-        type: 'string',
-        description: 'Inline citation format',
-      },
-      bibliographyVersion: {
-        type: 'string',
-        description: 'Bibliography entry format',
-      },
-    },
-  },
-};
-
-/**
  * Handles worldbuilding document creation
  */
-async function handleWorldbuildingDocument(parameters: any): Promise<ExecuteToolResponseSchema> {
-  const { worldName, worldType = 'other', keyElements = [], description } = parameters;
-  
-  // This is where we'd integrate with Claude for advanced generation
-  // For now, we'll implement a template-based approach
-  
-  // Default sections for worldbuilding
-  const defaultSections = [
-    'Overview',
-    'History',
-    'Geography',
-    'Culture',
-    'Magic/Technology',
-    'Political Structure',
-    'Key Locations',
-    'Notable Characters',
-    'Conflicts',
-    'Themes',
-  ];
-  
-  // Add any additional sections from keyElements
-  const additionalSections = keyElements.filter(elem => !defaultSections.includes(elem));
-  const allSections = [...defaultSections, ...additionalSections];
-  
-  // Create the LaTeX document
-  let latex = `\\documentclass[12pt,letterpaper]{article}
-\\usepackage{geometry}
-\\usepackage{titlesec}
-\\usepackage{xcolor}
-\\usepackage{graphicx}
-\\usepackage{hyperref}
-\\usepackage{fancyhdr}
-\\usepackage{booktabs}
-\\usepackage{tikz}
-\\usepackage{enumitem}
-
-% Speculative Fiction Worldbuilding Document Style
-\\geometry{margin=1in}
-\\hypersetup{colorlinks=true, linkcolor=blue, filecolor=magenta, urlcolor=cyan}
-\\pagestyle{fancy}
-\\fancyhf{}
-\\rhead{\\textit{${worldName}}}
-\\lhead{\\textit{${worldType} World}}
-\\cfoot{\\thepage}
-
-\\titleformat{\\section}{\\normalfont\\Large\\bfseries}{\\thesection}{1em}{}[{\\titlerule[0.8pt]}]
-\\titleformat{\\subsection}{\\normalfont\\large\\bfseries}{\\thesubsection}{1em}{}
-
-\\begin{document}
-
-\\begin{titlepage}
-  \\centering
-  \\vspace*{2cm}
-  {\\LARGE\\bfseries ${worldName} \\par}
-  \\vspace{1cm}
-  {\\Large A ${worldType} World\\par}
-  \\vspace{2cm}
-  
-  \\begin{abstract}
-  ${description}
-  \\end{abstract}
-  
-  \\vfill
-  {\\large \\today\\par}
-\\end{titlepage}
-
-\\tableofcontents
-\\newpage
-`;
-
-  // Add sections
-  allSections.forEach(section => {
-    latex += `\\section{${section}}
-This section contains details about the ${section.toLowerCase()} of ${worldName}.
-
-`;
-
-    // Add subsections for certain main sections
-    if (section === 'Geography') {
-      latex += `\\subsection{Physical Features}
-Describe the major physical features of the world here.
-
-\\subsection{Climate}
-Describe the climate patterns of the world here.
-
-\\subsection{Map}
-\\begin{center}
-[Insert map here]
-\\end{center}
-
-`;
-    } else if (section === 'Magic/Technology') {
-      latex += `\\subsection{Basic Principles}
-Outline the fundamental rules or principles of the ${worldType === 'fantasy' ? 'magic system' : 'technology'}.
-
-\\subsection{Limitations}
-Describe the limitations and costs of using ${worldType === 'fantasy' ? 'magic' : 'technology'} in this world.
-
-\\subsection{Cultural Impact}
-Explain how ${worldType === 'fantasy' ? 'magic' : 'technology'} has shaped the cultures of this world.
-
-`;
-    } else if (section === 'Notable Characters') {
-      latex += `\\subsection{Character Archetypes}
-Describe common character archetypes found in this world.
-
-\\subsection{Key Figures}
-List and describe important figures in the world's history or current events.
-
-`;
-    }
-  });
-  
-  // Add appendices for additional worldbuilding elements
-  latex += `\\appendix
-\\section{Glossary}
-\\begin{description}[style=nextline]
-\\item[Term] Definition of the term.
-\\item[Another Term] Definition of another term.
-\\end{description}
-
-\\section{Timeline of Key Events}
-\\begin{tabular}{p{3cm}p{12cm}}
-\\toprule
-\\textbf{Date/Time} & \\textbf{Event} \\\\
-\\midrule
-Time period & Description of historical event \\\\
-Another period & Description of another historical event \\\\
-\\bottomrule
-\\end{tabular}
-
-\\end{document}`;
-
-  return {
-    result: {
-      latex,
-      sections: allSections,
-    }
-  };
-}
 
 /**
  * Handles philosophical argument structure creation
@@ -734,183 +431,6 @@ Detail how this position contradicts or undermines the thesis.
 /**
  * Handles character profile creation
  */
-async function handleCharacterProfile(parameters: any): Promise<ExecuteToolResponseSchema> {
-  const { 
-    characterName, 
-    characterType = 'protagonist', 
-    basicDescription, 
-    worldContext = '',
-    detailLevel = 'detailed'
-  } = parameters;
-  
-  // Define sections based on detail level
-  let sections: string[] = [];
-  
-  // Basic sections for all detail levels
-  const basicSections = [
-    'Basic Information',
-    'Physical Description',
-    'Personality',
-    'Background',
-    'Motivations'
-  ];
-  
-  // Additional sections for detailed profile
-  const detailedSections = [
-    'Relationships',
-    'Skills & Abilities',
-    'Internal Conflicts',
-    'External Conflicts',
-    'Character Arc'
-  ];
-  
-  // More sections for comprehensive profile
-  const comprehensiveSections = [
-    'Psychological Profile',
-    'Cultural Context',
-    'Symbolic Significance',
-    'Alternative Interpretations',
-    'Development Notes'
-  ];
-  
-  // Build section list based on detail level
-  if (detailLevel === 'basic') {
-    sections = basicSections;
-  } else if (detailLevel === 'detailed') {
-    sections = [...basicSections, ...detailedSections];
-  } else { // comprehensive
-    sections = [...basicSections, ...detailedSections, ...comprehensiveSections];
-  }
-  
-  // Create LaTeX document
-  let latex = `\\documentclass[12pt]{article}
-\\usepackage{geometry}
-\\usepackage{titlesec}
-\\usepackage{xcolor}
-\\usepackage{tikz}
-\\usepackage{enumitem}
-\\usepackage{graphicx}
-\\usepackage{booktabs}
-\\usepackage{fancyhdr}
-\\usepackage{hyperref}
-
-% Character Profile Style
-\\geometry{margin=1in}
-\\hypersetup{colorlinks=true, linkcolor=blue, filecolor=magenta, urlcolor=cyan}
-\\pagestyle{fancy}
-\\fancyhf{}
-\\rhead{${characterName}}
-\\lhead{${characterType}}
-\\cfoot{\\thepage}
-
-\\titleformat{\\section}{\\normalfont\\Large\\bfseries}{\\thesection}{1em}{}[{\\titlerule[0.8pt]}]
-\\titleformat{\\subsection}{\\normalfont\\large\\bfseries}{\\thesubsection}{1em}{}
-
-\\begin{document}
-
-\\begin{center}
-{\\LARGE\\bfseries ${characterName}}\\\\[0.5cm]
-{\\large Character Profile - ${characterType}}\\\\[1cm]
-\\end{center}
-
-% Character summary
-\\begin{quotation}
-${basicDescription}
-\\end{quotation}
-
-`;
-
-  // Add world context if provided
-  if (worldContext) {
-    latex += `\\section*{World Context}
-${worldContext}
-
-`;
-  }
-
-  // Add content for each section
-  sections.forEach(section => {
-    latex += `\\section{${section}}\n`;
-    
-    // Add specific content for some sections
-    if (section === 'Basic Information') {
-      latex += `\\begin{tabular}{p{4cm}p{10cm}}
-\\textbf{Full Name:} & ${characterName} \\\\
-\\textbf{Role:} & ${characterType} \\\\
-\\textbf{Age:} & \\\\
-\\textbf{Occupation:} & \\\\
-\\textbf{Residence:} & \\\\
-\\end{tabular}
-
-`;
-    } else if (section === 'Skills & Abilities') {
-      latex += `\\subsection{Core Competencies}
-List the character's main skills here.
-
-\\subsection{Special Abilities}
-Describe any special powers or unique abilities.
-
-\\subsection{Weaknesses}
-Describe the character's key weaknesses or limitations.
-
-`;
-    } else if (section === 'Character Arc') {
-      latex += `\\subsection{Starting Point}
-Describe the character's state at the beginning.
-
-\\subsection{Key Transformations}
-List the major changes or growth points.
-
-\\subsection{End State}
-Describe the character's state at the end of their arc.
-
-`;
-    } else if (section === 'Symbolic Significance' && detailLevel === 'comprehensive') {
-      latex += `\\subsection{Archetypes}
-Discuss any archetypes this character represents.
-
-\\subsection{Symbolic Elements}
-Detail symbolic aspects of the character.
-
-\\subsection{Thematic Representation}
-Explain how the character embodies key themes.
-
-`;
-    } else {
-      // Default content placeholder for other sections
-      latex += `% Content for ${section}\n\n`;
-    }
-  });
-  
-  // Add visual relationship map for detailed and comprehensive profiles
-  if (detailLevel !== 'basic') {
-    latex += `\\section{Relationship Map}
-\\begin{center}
-\\begin{tikzpicture}
-\\node[draw, circle, fill=blue!20] (main) at (0,0) {${characterName}};
-\\node[draw, circle, fill=green!20] (char1) at (2,1) {Character 1};
-\\node[draw, circle, fill=green!20] (char2) at (-2,1) {Character 2};
-\\node[draw, circle, fill=orange!20] (char3) at (0,-2) {Character 3};
-
-\\draw (main) -- (char1) node[midway, above] {relationship};
-\\draw (main) -- (char2) node[midway, above] {relationship};
-\\draw (main) -- (char3) node[midway, right] {relationship};
-\\end{tikzpicture}
-\\end{center}
-
-`;
-  }
-  
-  // End document
-  latex += `\\end{document}`;
-
-  return {
-    result: {
-      latex,
-      sections,
-    }
-  };
-}
 
 /**
  * Handles concept map creation
@@ -1638,176 +1158,6 @@ ${events.length > 0 ?
 /**
  * Handles philosophical citation formatting
  */
-async function handlePhilosophicalCitation(parameters: any): Promise<ExecuteToolResponseSchema> {
-  const { 
-    sourceType, 
-    citationDetails,
-    citationStyle = 'chicago',
-    inlineCitation = true
-  } = parameters;
-  
-  // Extract citation details with defaults
-  const {
-    author = 'Author Name',
-    title = 'Title of Work',
-    year = '2023',
-    publisher = 'Publisher Name',
-    translator = '',
-    editor = '',
-    journal = '',
-    volume = '',
-    pages = '',
-    edition = '',
-    url = ''
-  } = citationDetails;
-  
-  // Generate citation based on style and source type
-  let inlineVersion = '';
-  let bibliographyVersion = '';
-  
-  if (citationStyle === 'chicago') {
-    // Chicago style citations
-    if (sourceType === 'book') {
-      inlineVersion = `(${author} ${year})`;
-      bibliographyVersion = `${author}. \\textit{${title}}${translator ? `, translated by ${translator}` : ''}${editor ? `, edited by ${editor}` : ''}${edition ? `, ${edition} ed.` : ''}. ${publisher}, ${year}.`;
-    } else if (sourceType === 'article') {
-      inlineVersion = `(${author} ${year})`;
-      bibliographyVersion = `${author}. ``${title}.'' \\textit{${journal}}${volume ? ` ${volume}` : ''}${pages ? `: ${pages}` : ''}, ${year}.`;
-    } else if (sourceType === 'primary-text') {
-      // For philosophical primary texts, often cited by section rather than page
-      inlineVersion = `(${author}, ${title.split(' ')[0]}, ${pages || 'section'})`;
-      bibliographyVersion = `${author}. \\textit{${title}}${translator ? `, translated by ${translator}` : ''}${editor ? `, edited by ${editor}` : ''}. ${publisher}, ${year}.`;
-    } else {
-      // Default format
-      inlineVersion = `(${author} ${year})`;
-      bibliographyVersion = `${author}. \\textit{${title}}. ${publisher}, ${year}.`;
-    }
-  } else if (citationStyle === 'apa') {
-    // APA style citations
-    if (sourceType === 'book') {
-      inlineVersion = `(${author}, ${year})`;
-      bibliographyVersion = `${author}. (${year}). \\textit{${title}}${edition ? ` (${edition} ed.)` : ''}${translator ? ` (${translator}, Trans.)` : ''}. ${publisher}.`;
-    } else if (sourceType === 'article') {
-      inlineVersion = `(${author}, ${year})`;
-      bibliographyVersion = `${author}. (${year}). ${title}. \\textit{${journal}}${volume ? `, ${volume}` : ''}${pages ? `, ${pages}` : ''}.`;
-    } else {
-      // Default format
-      inlineVersion = `(${author}, ${year})`;
-      bibliographyVersion = `${author}. (${year}). \\textit{${title}}. ${publisher}.`;
-    }
-  } else {
-    // Default to MLA style
-    if (sourceType === 'book') {
-      inlineVersion = `(${author} ${pages})`;
-      bibliographyVersion = `${author}. \\textit{${title}}${translator ? `, translated by ${translator}` : ''}${editor ? `, edited by ${editor}` : ''}. ${publisher}, ${year}.`;
-    } else if (sourceType === 'article') {
-      inlineVersion = `(${author} ${pages})`;
-      bibliographyVersion = `${author}. ``${title}.'' \\textit{${journal}}${volume ? `, vol. ${volume}` : ''}${pages ? `, ${pages}` : ''}, ${year}.`;
-    } else {
-      // Default format
-      inlineVersion = `(${author} ${pages || year})`;
-      bibliographyVersion = `${author}. \\textit{${title}}. ${publisher}, ${year}.`;
-    }
-  }
-  
-  // Generate BibTeX entry
-  let bibtexEntry = '';
-  const authorLastName = author.split(' ').pop() || 'Author';
-  const citeKey = `${authorLastName.toLowerCase()}${year}`;
-  
-  if (sourceType === 'book') {
-    bibtexEntry = `@book{${citeKey},
-  author = {${author}},
-  title = {${title}},
-  year = {${year}},
-  publisher = {${publisher}}${edition ? `,\n  edition = {${edition}}` : ''}${translator ? `,\n  translator = {${translator}}` : ''}${editor ? `,\n  editor = {${editor}}` : ''}
-}`;
-  } else if (sourceType === 'article') {
-    bibtexEntry = `@article{${citeKey},
-  author = {${author}},
-  title = {${title}},
-  journal = {${journal}}${volume ? `,\n  volume = {${volume}}` : ''}${pages ? `,\n  pages = {${pages}}` : ''},
-  year = {${year}}
-}`;
-  } else if (sourceType === 'primary-text') {
-    bibtexEntry = `@book{${citeKey},
-  author = {${author}},
-  title = {${title}},
-  year = {${year}},
-  publisher = {${publisher}}${translator ? `,\n  translator = {${translator}}` : ''}${editor ? `,\n  editor = {${editor}}` : ''}
-}`;
-  } else {
-    bibtexEntry = `@misc{${citeKey},
-  author = {${author}},
-  title = {${title}},
-  year = {${year}}${publisher ? `,\n  publisher = {${publisher}}` : ''}${url ? `,\n  url = {${url}}` : ''}
-}`;
-  }
-  
-  // Generate LaTeX document with citation examples
-  const latex = `\\documentclass{article}
-\\usepackage{geometry}
-\\usepackage{xcolor}
-\\usepackage{hyperref}
-\\usepackage{natbib}  % For citation management
-
-\\title{Philosophical Citation Example}
-\\author{}
-\\date{}
-
-\\begin{document}
-
-\\maketitle
-
-\\section{Citation Example}
-
-\\subsection{Source Information}
-\\begin{itemize}
-  \\item Source Type: ${sourceType}
-  \\item Author: ${author}
-  \\item Title: ${title}
-  \\item Year: ${year}
-  ${publisher ? `\\item Publisher: ${publisher}` : ''}
-  ${journal ? `\\item Journal: ${journal}` : ''}
-  ${volume ? `\\item Volume: ${volume}` : ''}
-  ${pages ? `\\item Pages: ${pages}` : ''}
-  ${translator ? `\\item Translator: ${translator}` : ''}
-  ${editor ? `\\item Editor: ${editor}` : ''}
-  ${edition ? `\\item Edition: ${edition}` : ''}
-  ${url ? `\\item URL: ${url}` : ''}
-\\end{itemize}
-
-\\subsection{${citationStyle.toUpperCase()} Style Citations}
-
-\\subsubsection{In-text Citation}
-${inlineCitation ? inlineVersion : 'No in-text citation requested.'}
-
-\\subsubsection{Bibliography Entry}
-${bibliographyVersion}
-
-\\subsection{BibTeX Entry}
-\\begin{verbatim}
-${bibtexEntry}
-\\end{verbatim}
-
-\\subsection{Usage Example}
-
-According to philosophical tradition ${inlineVersion}, the concept of being is fundamental to metaphysical inquiry. This perspective has been influential in continental philosophy.
-
-\\begin{thebibliography}{9}
-\\bibitem{${citeKey}} ${bibliographyVersion}
-\\end{thebibliography}
-
-\\end{document}`;
-
-  return {
-    result: {
-      latex,
-      inlineVersion,
-      bibliographyVersion
-    }
-  };
-}
 
 // Register implementation for each tool
 async function registerToolImplementations(server: MCPServer): Promise<void> {
